Make Host Layout responsive
Flexbox already takes care of most of the responsiveness so this just
brings the header font size in line with the other elements and applies
a minimum width to the left component and header.

The most common resolutions (720, 1080, 1440, 4K) are specifically
targeted to ensure they will look good on most desktops/laptops/tvs
with in between sizes taking a backseat at this time. That being said
the layout is functional from ~250px x ~250px up past 4k.