Add OptionList for the Host Settings Menu
OptionList is a subcomponent of the Host Settings Menu for rendering
accordion menus with dynamic text and functions controlled by state in
the Host Settings Menu.

The current implementation is for kicking
players, but it should be easy to extend the OptionList subcomponent to
other accordions in the settings menu with little if any modification.