Refactor HostSettingsMenu to accept more than one accordion
Refactor the OptionButton to receive separate enabled and disabled
onClick props rather than controlling execution of the onClick callback
inside of the OptionButton.

Refactor the OptionList to receive a "state" prop string to convert it
from a two way switch to a three way switch with "open", "enabled",
and "disabled" being the three states.

Update tests in accordance with the above changes.