Refactor OptionList and PlayerKicker API
Refactor OptionList (and tests) to take a single onClick prop in place
of the dual onEnabledClick and onNotEnabledClick. Also revise the
onListItemClick prop to just onItemClick.

Refactor PlayerKicker to take an onClickActions object prop containing
the different possible states as keys and the functions to execute
based on each of those keys as the values. Select the onClick getting
passed into the OptionList using this object and the accordionState.

Revise HostSettingsMenu to match the new PlayerKicker API.