Lift lobby id generation from Lobby to LobbyList
Update lobby id / join code shuffling to match this change.

The map for join code shuffling now goes:
The message is receiving in Lobby from the host, the lobby then
executes a callback provided by LobbyList where the changing of the
id now happens. At the end of this callback LobbyList calls a Lobby
function with the new id. Lobby then updates its internal id and sends
the new join code message to the host.

Tests for Lobby were also updated as required.