Refactor HostSettingsMenu API
Rather than just taking an array of components in this commit has
HostSettingsMenu take an array of objects with the component and a
specified 'type'. This allows additional components to be added without
having the add an entry for them in the HostSettingsMenu switch (so
long as they are build using the accordion or button settings sub
components that are built so far).